Rename components/MenuList.vue to components/_shared/SideNavigation/MenuList.vue
